Add Twitter handler, and refactor how handlers handle redirection
The old callback-stores-the-redirection-target stuff was to support the bad
old days of signed URLs for email login, but we're not doing that anymore
since all the login handlers require maintaining state anyway. So now the
redirection target just gets managed by the handler as part of the transaction
state. This was necessary because Twitter requires an exact match on callback
URL, and there are probably other providers which will as well; the fact the
other providers could be loosy-goosy about the callback URL isn't really a
feature.

Also, this makes things more flexible for frameworks which aren't Flask, since
there are situations where simply putting the redirection target into the URL
might not be the appropriate choice.

This also aids privacy somewhat since now the identity server
won't know anything about what URL the user is trying to log into anyway.

Basically this makes things more sensible in the long run for everyone.